[util, regtool] Introduce vendor specific attributes
Vendor specific attributes are optional and does not affect reggen code
gen, they can be used by different tools for a vendor specific use case.
With this commit, vendor specific attributes can be defined either with
a argument or just by placing a file with the name
'vendor_specific_fiels.hjson' to the directory 'util/reggen'.